+++ src/types/log/log.h
@@ -6,7 +6,7 @@
 /*   By: mimeyer <mimeyer@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/05/29 23:48:40 by mimeyer           #+#    #+#             */
-/*   Updated: 2026/05/30 14:02:42 by mimeyer          ###   ########.fr       */
+/*   Updated: 2026/06/04 23:49:54 by mimeyer          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ typedef struct s_log
 }					t_log;
 
 t_log				*log_create(void);
-int					log_print(t_log *log, int time, int idx, char *arg,
+int					log_print(t_log *log, uint64_t time, int idx, char *arg,
 						int death, int finished);
 int					log_free(t_log *log);
