+++ utils/ft_gettimems.c
@@ -6,16 +6,16 @@
 /*   By: mimeyer <mimeyer@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/05/31 21:10:08 by mimeyer           #+#    #+#             */
-/*   Updated: 2026/05/31 21:33:49 by mimeyer          ###   ########.fr       */
+/*   Updated: 2026/06/04 23:45:54 by mimeyer          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "utils.h"
 
-size_t	gettimems(void)
+uint64_t	gettimems(void)
 {
 	struct timeval	tv;
 
 	gettimeofday(&tv, NULL);
-	return ((size_t)tv.tv_sec * 1000 + (size_t)tv.tv_usec / 1000);
+	return ((uint64_t)tv.tv_sec * 1000 + (uint64_t)tv.tv_usec / 1000);
 }
