+++ utils/utils.h
@@ -6,7 +6,7 @@
 /*   By: mimeyer <mimeyer@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/05/02 11:21:44 by mimeyer           #+#    #+#             */
-/*   Updated: 2026/05/31 21:17:46 by mimeyer          ###   ########.fr       */
+/*   Updated: 2026/06/04 23:53:21 by mimeyer          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,6 +19,7 @@
 # include <string.h>
 # include <sys/time.h>
 # include <unistd.h>
+# include <stdint.h>
 
 void	*ft_calloc(size_t nmemb, size_t size);
 size_t	ft_strlcpy(char *dest, const char *src, size_t n);
@@ -26,9 +27,9 @@ void	*ft_memcpy(void *dest, const void *src, size_t n);
 size_t	ft_strlen(const char *s);
 void	ft_putstr_fd(char *s, int fd);
 char	**append_msg(char **msg, char *arg);
-char	*ft_itoa(int n);
+char	*ft_itoa(int64_t n);
 int		ft_isdigit(int c);
 int		ft_isstr(char *arg, int (*check)(int));
-size_t	gettimems(void);
+uint64_t	gettimems(void);
 
 #endif
